fix: freeze test cards for current user
- each unique user must have a unique set of cards that should not change
- move service interfaces to related folder
